+++ types/location/struct.go
@@ -1,6 +1,6 @@
 package location
 
 type Location struct {
-	Latitude  *float32 `json:"lat"  validate:"required"`
-	Longitude *float32 `json:"long" validate:"required"`
+	Latitude  *float64 `json:"lat"  validate:"required"`
+	Longitude *float64 `json:"long" validate:"required"`
 }
